Extract footnotes scanner
