+++ src/scan/mod.rs
@@ -2,6 +2,7 @@
 //! and reports problems found as `Issues`.
 
 pub(crate) mod duplicate_sections;
+pub(crate) mod footnotes;
 pub(crate) mod section_capitalization;
 pub(crate) mod section_level;
 pub(crate) mod unordered_sections;
